Support no_std
