+++ .github/workflows/check.yml
@@ -44,8 +44,13 @@ jobs:
           target: riscv32imc-unknown-none-elf
           toolchain: stable
           components: rust-src
-      - name: Run tests no_std
-        run: cargo test --features=no_std
+      - name: Run no_std example
+        run: |
+          cargo fmt --all -- --check
+          cargo fix
+          cargo clippy --all -- -D warnings
+          cargo b --release
+        working-directory: ./is_debug_no_std
