+++ Cargo.toml
@@ -13,7 +13,8 @@ edition = "2018"
 path = "lib.rs"
 
 [features]
-no_std = []
+default = ["std"]
+std = []
 
 [dev-dependencies]
 xshell = "0.2.7"
