+++ is_debug_no_std/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "is_debug_no_std"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+esp-backtrace = { version = "0.14.2", features = [
+	"esp32c3",
+	"exception-handler",
+	"panic-handler",
+	"println",
+] }
+
+esp-hal = { version = "0.22.0", features = [
+	"esp32c3",
+] }
+esp-println = { version = "0.12.0", features = ["esp32c3", "log"] }
+log = { version = "0.4.22" }
+serde = { version = "*", default-features = false }
+is_debug = { path = "../", default-features = false }
