+++ is_debug_no_std/rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "stable"
+components = ["rust-src"]
+targets = ["riscv32imc-unknown-none-elf"]
