+++ is_debug_no_std/src/main.rs
@@ -1,0 +1,21 @@
+#![no_std]
+#![no_main]
+
+use esp_backtrace as _;
+use esp_hal::delay::Delay;
+use esp_hal::prelude::*;
+use log::info;
+
+#[entry]
+fn main() -> ! {
+    esp_println::logger::init_logger_from_env();
+
+    let delay = Delay::new();
+
+    info!("{}", is_debug::is_debug());
+
+    loop {
+        info!("Hello world!");
+        delay.delay(500.millis());
+    }
+}
